Select module import
+ similar fixes for completion
+ split tests into 23 and 25
  - tests for disabled module imports remain in CompletionTests23

Fixes https://github.com/eclipse-jdt/eclipse.jdt.core/issues/4371
